requirements.txt
tator~=1.0.0
python-dotenv
piexif
click~=8.1.3
deepsea-ai
numpy~=1.24.3
pillow~=9.5.0
tensorflow